% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsensate.R
\name{backward_solver}
\alias{backward_solver}
\title{Non-Parametric Binary Sensitivity Analysis}
\usage{
backward_solver(X, Y, R, fi, method, A, fi_change = FALSE)
}
\arguments{
\item{X}{The treatment variable (vector with 0/1 entries).}

\item{Y}{The outcome variable (vector with 0/1 entries).}

\item{R}{The set of proxies for the confounding variables Z (matrix with 0/1
entries).}

\item{fi}{Sensitivity parameters of the analysis. List with two entries, both
of which are lists. The entry \code{fi[[1+x]][[1+y]]} contains the sensitivity
parameter corresponding to the probability that a confounder is unobserved for
the specific value of X = x, Y = y.}

\item{A}{Sensitivity parameters that  List with two entries, both
of which are lists. The entry \code{A[[1+x]][[1+y]]} contains the entries of the
fidelity matrix A_xy for X = x, Y = y. Default to NULL (in which case) \code{fi} is
used to infer the value of A_xy.}
}
\description{
Non-Parametric Binary Sensitivity Analysis
}
