% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsensate.R
\name{binsensate}
\alias{binsensate}
\title{Binary Sensitivity Analysis}
\usage{
binsensate(X, Y, R, fi, solver = c("backward", "two-stage-em"), ...)
}
\arguments{
\item{X}{The treatment variable (vector with 0/1 entries).}

\item{Y}{The outcome variable (vector with 0/1 entries).}

\item{R}{The set of proxies for the confounding variables Z (matrix with 0/1
entries).}

\item{fi}{Sensitivity parameters of the analysis. List with two entries, both
of which are lists. The entry \code{fi[[1+x]][[1+y]]} contains the sensitivity
parameter corresponding to the probability that a confounder is unobserved for
the specific value of X = x, Y = y.}

\item{solver}{Character scalar describing the solution method,
either \code{"backward"} (non-parametric approach)
or \code{"two-stage-em"} (parametric EM).}

\item{...}{Further arguments passed to subroutine \code{\link{two_stage_em}}.}
}
\description{
Binary Sensitivity Analysis
}
