% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsensate.R
\name{em_solver_cts}
\alias{em_solver_cts}
\title{Binary Sensitivity Analysis using Expectation-Maximization in an Exponential
Family with Correctly Observed Continuous Confounders}
\usage{
em_solver_cts(
  X,
  Y,
  R,
  W,
  fi,
  method,
  A,
  mc_per_samp = 10,
  n_epoch = 10,
  rand_init = FALSE,
  se = FALSE,
  detect_viol = FALSE,
  mc_cores = NULL,
  tau_xw = NULL,
  tau_yw = NULL
)
}
\arguments{
\item{X}{The treatment variable (vector with 0/1 entries).}

\item{Y}{The outcome variable (vector with 0/1 entries).}

\item{R}{The set of proxies for the confounding variables Z (matrix with 0/1
entries).}

\item{W}{The set of correctly observed continuous confounders, which should be
mean centered and variance normalized. Defaults to \code{NULL}, meaning no continuous
confounders are included. If \code{W} is included, the continous EM routine is used
automatically (\code{W} cannot be used with the \code{\link{backward_solver}}).}

\item{fi}{Sensitivity parameters of the analysis. List with two entries, both
of which are lists. The entry \code{fi[[1+x]][[1+y]]} contains the sensitivity
parameter corresponding to the probability that a confounder is unobserved for
the specific value of X = x, Y = y.}

\item{method}{Character taking values \code{"IF"} for independent fidelity, or
\code{"ZINF"} for the zero-inflation fidelity pattern.}

\item{A}{Sensitivity parameters that  List with two entries, both
of which are lists. The entry \code{A[[1+x]][[1+y]]} contains the entries of the
fidelity matrix A_xy for X = x, Y = y. Default to NULL (in which case) \code{fi} is
used to infer the value of A_xy.}

\item{mc_per_samp}{Number of Monte Carlo samples drawn for each observed
sample in the data during the expectation-maximization algorithm.}

\item{n_epoch}{Number of epochs for the expectation-maximization algorithm.}

\item{rand_init}{Whether the expectation-maximization parameters should have
a random initialization. Defaults to \code{FALSE}.}

\item{se}{Logical indicating whether to compute standard errors using the
method of Louis. Defaults to \code{FALSE}.}

\item{detect_viol}{Logical indicating whether to test for ill-posedness of the
inverse problem. Defaults to \code{FALSE}.}

\item{mc_cores}{Number of cores used for parallelization during Monte Carlo
sampling in the M-step.}

\item{tau_xw}{A transformation function applied to the continuous confounders
W, after which the transformed data is included in the regression function
for the treatment variable X.}

\item{tau_yw}{A transformation function applied to the continuous confounders
W, after which the transformed data is included in the regression function
for the outcome variable Y.}
}
\description{
Binary Sensitivity Analysis using Expectation-Maximization in an Exponential
Family with Correctly Observed Continuous Confounders
}
