% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsensate.R
\name{two_stage_em}
\alias{two_stage_em}
\title{Binary Sensitivity Analysis using Expectation-Maximization in an Exponential
Family}
\usage{
two_stage_em(X, Y, R, fi, mc_per_samp = 10, n_epoch = 10, rand_init = FALSE)
}
\arguments{
\item{X}{The treatment variable (vector with 0/1 entries).}

\item{Y}{The outcome variable (vector with 0/1 entries).}

\item{R}{The set of proxies for the confounding variables Z (matrix with 0/1
entries).}

\item{fi}{Sensitivity parameters of the analysis. List with two entries, both
of which are lists. The entry \code{fi[[1+x]][[1+y]]} contains the sensitivity
parameter corresponding to the probability that a confounder is unobserved for
the specific value of X = x, Y = y.}

\item{mc_per_samp}{Number of Monte Carlo samples drawn for each observed
sample in the data during the expectation-maximization algorithm.}

\item{n_epoch}{Number of epochs for the expectation-maximization algorithm.}

\item{rand_init}{Whether the expectation-maximization parameters should have
a random initialization. Defaults to \code{FALSE}.}
}
\description{
Binary Sensitivity Analysis using Expectation-Maximization in an Exponential
Family
}
