% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsensate.R
\name{binsensate}
\alias{binsensate}
\title{Binary Sensitivity Analysis}
\usage{
binsensate(
  X,
  Y,
  R,
  fi,
  method = c("IF", "ZINF"),
  A = NULL,
  solver = c("backward", "em"),
  ...
)
}
\arguments{
\item{X}{The treatment variable (vector with 0/1 entries).}

\item{Y}{The outcome variable (vector with 0/1 entries).}

\item{R}{The set of proxies for the confounding variables Z (matrix with 0/1
entries).}

\item{fi}{Sensitivity parameters of the analysis. List with two entries, both
of which are lists. The entry \code{fi[[1+x]][[1+y]]} contains the sensitivity
parameter corresponding to the probability that a confounder is unobserved for
the specific value of X = x, Y = y.}

\item{method}{Character taking values \code{"IF"} for independent fidelity, or
\code{"ZINF"} for the zero-inflation fidelity pattern.}

\item{A}{Sensitivity parameters that  List with two entries, both
of which are lists. The entry \code{A[[1+x]][[1+y]]} contains the entries of the
fidelity matrix A_xy for X = x, Y = y. Default to NULL (in which case) \code{fi} is
used to infer the value of A_xy.}

\item{solver}{Character scalar describing the solution method,
either \code{"backward"} (non-parametric approach) or \code{"em"} (parametric EM).}

\item{...}{Further arguments passed to subroutine \code{\link{em_solver}}.}
}
\description{
Binary Sensitivity Analysis
}
